spending-alerts/requirements.txt
psycopg2-binary
pandas
sqlalchemy
scikit-learn
fastapi
uvicorn
